requirements.txt
# Network Pharmacology Toolkit - Dependencies
# Install with: pip install -r requirements.txt

# =============================================================================
# CORE
# =============================================================================
pyyaml>=6.0
click>=8.1.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0

# =============================================================================
# NETWORK ANALYSIS
# =============================================================================
networkx>=3.0
# python-igraph>=0.10.0  # Optional, for advanced graph algorithms

# =============================================================================
# API & WEB SCRAPING
# =============================================================================
requests>=2.28.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0  # Async HTTP requests
lxml>=4.9.0

# =============================================================================
# CHEMISTRY & BIOINFORMATICS
# =============================================================================
pubchempy>=1.0.4
biopython>=1.81

# =============================================================================
# ENRICHMENT ANALYSIS
# =============================================================================
gseapy>=1.0.0

# =============================================================================
# VISUALIZATION
# =============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
matplotlib-venn>=0.11.0
pyvis>=0.3.0  # Interactive network visualization

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv>=1.0.0
rich>=13.0.0  # Pretty console output
tabulate>=0.9.0
openpyxl>=3.1.0  # Excel export

# =============================================================================
# TESTING (Optional)
# =============================================================================
pytest>=7.0.0
pytest-cov>=4.0.0
